+++ tests/test_all.py
@@ -7,7 +7,7 @@ import test_noplot
 import test_kyoko
 import test_koroshi
 import test_yunow
-
+import test_pants
 
 def suite():
     '''Packing all tests.
@@ -24,6 +24,7 @@ def suite():
         unittest.makeSuite(test_noplot.StoryTest),
         unittest.makeSuite(test_koroshi.StoryTest),
         unittest.makeSuite(test_yunow.StoryTest),
+        unittest.makeSuite(test_pants.StoryTest),
         # novema
         # cobalt
         unittest.makeSuite(test_kyoko.StoryTest),